% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLATCHE3_export.R
\name{export}
\alias{export}
\alias{export.default}
\alias{export.SPLATCHE3_settings}
\alias{export.SPLATCHE3_PopDensity}
\alias{export.SPLATCHE3_Veg2dyn}
\alias{export.SPLATCHE3_Veg2phase}
\alias{export.SPLATCHE3_ArrivalCell}
\alias{export.SPLATCHE3_Sample}
\alias{export.SPLATCHE3_Genetic}
\title{write \strong{SPLATCHE3_...} to file}
\usage{
export(x, tmpdir, ...)

\method{export}{default}(x, tmpdir, ...)

\method{export}{SPLATCHE3_settings}(x, tmpdir, ...)

\method{export}{SPLATCHE3_PopDensity}(x, tmpdir, ...)

\method{export}{SPLATCHE3_Veg2dyn}(x, tmpdir, ...)

\method{export}{SPLATCHE3_Veg2phase}(x, tmpdir, ...)

\method{export}{SPLATCHE3_ArrivalCell}(x, tmpdir, ...)

\method{export}{SPLATCHE3_Sample}(x, tmpdir, ...)

\method{export}{SPLATCHE3_Genetic}(x, tmpdir, ...)
}
\arguments{
\item{x}{an object of class SPLATCHE3_settings}

\item{tmpdir}{a non-empty character vector giving a directory name}

\item{...}{ignored}
}
\description{
write \strong{SPLATCHE3_...} to file
}
