% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLATCHE3_install.R
\name{install}
\alias{install}
\title{quick setup of SPLATCHE3}
\usage{
install(path = "~/splatche3", os = "Linux")
}
\arguments{
\item{path}{character. Where the executable should be stored}

\item{os}{character. Operating system ("Linux", "Windows", "Mac")}
}
\value{
character. Path to executable
}
\description{
quick setup of SPLATCHE3
}
