% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLATCHE3_basic.R
\name{SPLATCHE3_basic}
\alias{SPLATCHE3_basic}
\alias{as.SPLATCHE3_settings}
\alias{is.SPLATCHE3_settings}
\alias{as.SPLATCHE3_PopDensity}
\alias{is.SPLATCHE3_PopDensity}
\title{class functions}
\usage{
as.SPLATCHE3_settings(x, ...)

is.SPLATCHE3_settings(x, ...)

as.SPLATCHE3_PopDensity(x, ...)

is.SPLATCHE3_PopDensity(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{ignored}
}
\description{
class functions
}
