% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLATCHE3_settings_write.R
\name{write_settings}
\alias{write_settings}
\alias{write_settings.default}
\alias{write_settings.SPLATCHE3_settings}
\title{write \strong{SPLATCHE3_settings}s to files}
\usage{
write_settings(x, con = tempfile(), ...)

\method{write_settings}{default}(x, con = tempfile(), ...)

\method{write_settings}{SPLATCHE3_settings}(x, con = tempfile(), ...)
}
\arguments{
\item{x}{an object of class SPLATCHE3_settings}

\item{con}{a connection object or a character string}

\item{...}{ignored}
}
\description{
write \strong{SPLATCHE3_settings}s to files
}
\examples{
#write_settings(
#   example_SPLATCHE3_settings
#)

}
