% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLATCHE3_export.R
\name{export}
\alias{export}
\alias{export.default}
\alias{export.SPLATCHE3_settings}
\alias{export.SPLATCHE3_PopDensity}
\title{write \strong{SPLATCHE3_settings} to file}
\usage{
export(x, con = tempfile(), ...)

\method{export}{default}(x, con = tempfile(), ...)

\method{export}{SPLATCHE3_settings}(x, con = tempfile(), ...)

\method{export}{SPLATCHE3_PopDensity}(x, con = tempfile(), ...)
}
\arguments{
\item{x}{an object of class SPLATCHE3_settings}

\item{con}{a connection object or a character string}

\item{...}{ignored}
}
\description{
write \strong{SPLATCHE3_settings} to file
}
