% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLATCHE3_export.R
\name{export}
\alias{export}
\alias{export.default}
\alias{export.SPLATCHE3_settings}
\alias{export.SPLATCHE3_PopDensity}
\alias{export.SPLATCHE3_Veg2dyn}
\alias{export.SPLATCHE3_Veg2phase}
\alias{export.SPLATCHE3_ArrivalCell}
\alias{export.SPLATCHE3_Sample}
\alias{export.SPLATCHE3_Genetic}
\title{export SPLATCHE3 settings files from R objects}
\usage{
export(x, filename, outdir, ...)

\method{export}{default}(x, filename, outdir, ...)

\method{export}{SPLATCHE3_settings}(x, filename = "settings.txt", outdir, ...)

\method{export}{SPLATCHE3_PopDensity}(x, filename, outdir, ...)

\method{export}{SPLATCHE3_Veg2dyn}(x, filename, outdir, ...)

\method{export}{SPLATCHE3_Veg2phase}(x, filename, outdir, ...)

\method{export}{SPLATCHE3_ArrivalCell}(x, filename, outdir, ...)

\method{export}{SPLATCHE3_Sample}(x, filename, outdir, ...)

\method{export}{SPLATCHE3_Genetic}(x, filename, outdir, ...)
}
\arguments{
\item{x}{object of class SPLATCHE3_settings}

\item{filename}{character. Export file name}

\item{outdir}{character. Path to output directory}

\item{...}{ignored}
}
\description{
export SPLATCHE3 settings files from R objects
}
