% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLATCHE3_run.R
\name{run}
\alias{run}
\title{run SPLATCHE3}
\usage{
run(settings, tmpdir = "~/test/splatche", exe = "~/splatche3/splatche3")
}
\arguments{
\item{settings}{object of class SPLATCHE3_settings}

\item{tmpdir}{character. Path to output_dir directory}

\item{exe}{character. Path to SPLATCHE3 executable}
}
\value{

}
\description{
run SPLATCHE3
}
