% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLATCHE3_run.R
\name{run}
\alias{run}
\title{run SPLATCHE3}
\usage{
run(
  settings,
  output = file.path(tempdir(), "splatche3"),
  exe = "~/splatche3/splatche3"
)
}
\arguments{
\item{settings}{object of class SPLATCHE3_settings}

\item{output}{character. Path to output directory}

\item{exe}{character. Path to SPLATCHE3 executable}
}
\value{

}
\description{
run SPLATCHE3
}
