% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLATCHE3_basic.R
\name{as.SPLATCHE3_settings}
\alias{as.SPLATCHE3_settings}
\alias{is.SPLATCHE3_settings}
\alias{as.SPLATCHE3_PopDensity}
\alias{is.SPLATCHE3_PopDensity}
\alias{as.SPLATCHE3_Veg2dyn}
\alias{is.SPLATCHE3_Veg2dyn}
\alias{as.SPLATCHE3_Veg2phase}
\alias{is.SPLATCHE3_Veg2phase}
\alias{as.SPLATCHE3_ArrivalCell}
\alias{is.SPLATCHE3_ArrivalCell}
\alias{as.SPLATCHE3_Sample}
\alias{is.SPLATCHE3_Sample}
\alias{as.SPLATCHE3_Genetic}
\alias{is.SPLATCHE3_Genetic}
\title{SPLATCH3 settings objects}
\usage{
as.SPLATCHE3_settings(x, ...)

is.SPLATCHE3_settings(x, ...)

as.SPLATCHE3_PopDensity(x, ...)

is.SPLATCHE3_PopDensity(x, ...)

as.SPLATCHE3_Veg2dyn(x, ...)

is.SPLATCHE3_Veg2dyn(x, ...)

as.SPLATCHE3_Veg2phase(x, ...)

is.SPLATCHE3_Veg2phase(x, ...)

as.SPLATCHE3_ArrivalCell(x, ...)

is.SPLATCHE3_ArrivalCell(x, ...)

as.SPLATCHE3_Sample(x, ...)

is.SPLATCHE3_Sample(x, ...)

as.SPLATCHE3_Genetic(x, ...)

is.SPLATCHE3_Genetic(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
SPLATCH3 settings objects
}
